Got the new select bar mostly working
The change block function needs to be changed so it assigns the block info. Additionally, the placing needs to be done using an active block info instead of an active block as this will allow for the usage of the other sprites.

Finally, the folding and unfolding needs to be sorted.